+++ src/Transformer/NullableTransformer.php
@@ -25,6 +25,11 @@ final readonly class NullableTransformer implements \Stringable, TransformerInte
     ) {
     }
 
+    public function getItemTransformer(): TransformerInterface
+    {
+        return $this->itemTransformer;
+    }
+
     public function transform(Expr $input, Expr $target, PropertyMetadata $propertyMapping, UniqueVariableScope $uniqueVariableScope, Expr $source, ?Expr $existingValue = null): array
     {
         [$output, $itemStatements] = $this->itemTransformer->transform($input, $target, $propertyMapping, $uniqueVariableScope, $source, $existingValue);